refactored validation
