--- helpers/getById.js
@@ -1,21 +1,0 @@
-const contacts = require('../models/contacts')
-const HttpError = require('./HttpError')
-
-const getById = async (req, res, next) => {
-  
-  try {
-    const { contactId } = req.params
-    const result = await contacts.getContactById(contactId)
-    console.log(result)
-    if (!result) {
-      throw HttpError(404, "Not found")
-      
-    }
-    res.json(result)
-  } catch (error) {
-    next(error)
-    
-  }
-}
-
-module.exports = getById
