+++ helpers/index.js
@@ -1,15 +1,7 @@
 const HttpError = require("./HttpError")
-const getContacts = require("./getContacts")
-const deleteById = require("./deleteById")
-const getById = require("./getById")
-const postContact = require("./postContact")
-const putContact = require("./putContact")
+const ctrlWrapper = require("./ctrlWrapper")
 
 module.exports = {
   HttpError,
-  getContacts,
-  deleteById,
-  getById,
-  postContact,
-  putContact,
+  ctrlWrapper  
 }
